Moved MavenProject into maven package
