+++ spring-rewrite-commons-launcher/src/test/java/org/springframework/rewrite/parser/maven/MavenProjectSorterTest.java
@@ -26,7 +26,6 @@ import org.junit.jupiter.api.io.TempDir;
 import org.mockito.Mockito;
 import org.openrewrite.maven.utilities.MavenArtifactDownloader;
 import org.springframework.core.io.Resource;
-import org.springframework.rewrite.parser.MavenProject;
 import org.springframework.rewrite.test.util.DummyResource;
 import org.springframework.rewrite.utils.LinuxWindowsPathUnifier;
 import org.springframework.rewrite.utils.ResourceUtil;
